list group added to listGroup tsx file
