+++ src/App.tsx
@@ -1,19 +1,7 @@
+import ListGroup from "./components/ListGroup";
+
 function App() {
-  let items = ["New York", "San Francisco", "Tokyo", "Germany", "Paris"];
-  // items = [];
-  return (
-    <>
-      <h1>List</h1>
-      {items.length === 0 && <p>No items found</p>}
-      <ul className="list-group">
-        {items.map((item) => (
-          <li key={item} className="list-group-item">
-            {item}
-          </li>
-        ))}
-      </ul>
-    </>
-  );
+  return <ListGroup></ListGroup>;
 }
 
 export default App;
