+++ src/components/ListGroup.tsx
@@ -1,5 +1,19 @@
-function ListGroup(){
-    return <h1>List Group</h1>;
+function ListGroup() {
+  let items = ["New York", "San Francisco", "Tokyo", "Germany", "Paris"];
+  // items = [];
+  return (
+    <>
+      <h1>List</h1>
+      {items.length === 0 && <p>No items found</p>}
+      <ul className="list-group">
+        {items.map((item) => (
+          <li key={item} className="list-group-item">
+            {item}
+          </li>
+        ))}
+      </ul>
+    </>
+  );
 }
 
 export default ListGroup;
